preprocessing/requirements.txt
fastapi==0.100.0
uvicorn==0.23.2
python-multipart==0.0.6
redis==4.6.0
numba
opencv-python-headless
pytesseract
pillow
rembg
gdown
matplotlib
onnxruntime